# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ApiErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  handlers.CreateArtistRequest:
    properties:
      cover_url:
        example: https://example.com/artist.jpg
        type: string
      name:
        example: The Beatles
        type: string
      spotify_id:
        example: 6ml0jHmy7SNFWckrZblO5B
        type: string
    required:
    - name
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  handlers.UpdateArtistRequest:
    properties:
      cover_url:
        example: https://example.com/artist.jpg
        type: string
      name:
        example: The Beatles
        type: string
      spotify_id:
        example: 6ml0jHmy7SNFWckrZblO5B
        type: string
    type: object
...
  models.Album:
    properties:
      artist_id:
        type: integer
      comment:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      listened:
        type: boolean
      listened_at:
        type: string
      rating:
        type: integer
      spotify_id:
        type: string
      title:
        type: string
      year:
        type: integer
    required:
    - artist_id
    - created_at
    - id
    - listened
    - title
    type: object
  models.Artist:
    properties:
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      spotify_id:
        type: string
    required:
    - created_at
    - id
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: A music management API with artists and albums
  title: Better Music API
  version: "1.0"
paths:
...
  /api/artists:
    get:
      description: Get all artists for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Artist'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get all artists
      tags:
      - artists
    post:
      consumes:
      - application/json
      description: Create a new artist for the authenticated user
      parameters:
      - description: Artist data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateArtistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Artist'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Artist already exists
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Create a new artist
      tags:
      - artists
  /api/artists/{id}:
    delete:
      description: Delete an artist by ID (must have no albums)
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Artist has albums
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Delete an artist
      tags:
      - artists
    get:
      description: Get a specific artist by ID for the authenticated user
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Artist'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get a specific artist
      tags:
      - artists
    put:
      consumes:
      - application/json
      description: Update an artist's name and/or Spotify ID
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateArtistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Invalid request or no fields provided
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "409":
          description: Artist already exists
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Update an artist
      tags:
      - artists
  /api/artists/{id}/albums:
    get:
      description: Get all albums for a specific artist
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Album'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "404":
          description: Artist not found
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/handlers.ApiErrorResponse'
      security:
      - Bearer: []
      summary: Get all albums by an artist
      tags:
      - artists
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"